Refactor web app into package
